csv-agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.37.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
cerebras-cloud-sdk>=1.50.0
python-dotenv>=1.0.0
python-multipart>=0.0.19
